+++ src/transformers/models/wav2vec2/feature_extraction_wav2vec2.py
@@ -20,7 +20,7 @@ from typing import List, Optional, Union
 
 import numpy as np
 
-from ...feature_extraction_utils import BatchFeature, PreTrainedFeatureExtractor
+from ...feature_extraction_sequence_utils import BatchSequenceFeature, PreTrainedSequenceFeatureExtractor
 from ...file_utils import PaddingStrategy, TensorType
 from ...utils import logging
 
@@ -28,7 +28,7 @@ from ...utils import logging
 logger = logging.get_logger(__name__)
 
 
-class Wav2Vec2FeatureExtractor(PreTrainedFeatureExtractor):
+class Wav2Vec2FeatureExtractor(PreTrainedSequenceFeatureExtractor):
     r"""
     Constructs a Wav2Vec2 feature extractor.
 
@@ -93,7 +93,7 @@ class Wav2Vec2FeatureExtractor(PreTrainedFeatureExtractor):
         return_tensors: Optional[Union[str, TensorType]] = None,
         sampling_rate: Optional[int] = None,
         **kwargs
-    ) -> BatchFeature:
+    ) -> BatchSequenceFeature:
         """
         Main method to featurize and prepare for the model one or several sequence(s). sequences.
 
@@ -179,7 +179,7 @@ class Wav2Vec2FeatureExtractor(PreTrainedFeatureExtractor):
             raw_speech = self.zero_mean_unit_var_norm(raw_speech)
 
         # convert into correct format for padding
-        encoded_inputs = BatchFeature({"input_values": raw_speech})
+        encoded_inputs = BatchSequenceFeature({"input_values": raw_speech})
 
         padded_inputs = self.pad(
             encoded_inputs,
